+++ src/common/providers/reference_resolver.py
@@ -2,8 +2,8 @@ from typing import Callable
 
 from common.address import Address
 from common.exceptions import ApplicationException
+from common.providers.address_resolver import ResolvedAddress, resolve_address
 from common.utils.data_structure.is_reference import is_reference
-from common.utils.resolve_address import ResolvedAddress, resolve_address
 from storage.data_source_class import DataSource
