+++ src/common/providers/storage_recipe_provider.py
