+++ src/features/attribute/use_cases/get_attribute_use_case.py
@@ -1,7 +1,7 @@
 from authentication.models import User
 from common.address import Address
 from common.tree.tree_node import Node
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 from storage.internal.data_source_repository import get_data_source
