+++ src/features/document/document_feature.py
@@ -7,10 +7,10 @@ from pydantic import Json, conint
 from authentication.authentication import auth_w_jwt_or_pat
 from authentication.models import User
 from common.address import Address
+from common.providers.blueprint_provider import get_blueprint_provider
+from common.providers.storage_recipe_provider import storage_recipe_provider
 from common.responses import create_response, responses
-from common.utils.blueprint_provider import get_blueprint_provider
-from common.utils.get_storage_recipe import storage_recipe_provider
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 from storage.internal.data_source_repository import get_data_source
 
 from .use_cases.add_document_use_case import add_document_use_case
