+++ src/features/document/use_cases/check_exsistence_use_case.py
@@ -1,6 +1,6 @@
 from common.address import Address
 from common.exceptions import NotFoundException
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 
 
 def check_existence_use_case(address: Address, document_service: DocumentService) -> bool:
