+++ src/features/entity/use_cases/instantiate_entity.py
@@ -1,7 +1,7 @@
 from authentication.models import User
 from common.entity.create_entity import CreateEntity
 from restful.request_types.shared import Entity
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 
 
 def instantiate_entity_use_case(basic_entity: Entity, user: User) -> dict:
