+++ src/features/entity/use_cases/validate_entity.py
@@ -1,7 +1,7 @@
 from authentication.models import User
 from common.entity.validators import validate_entity, validate_entity_against_self
 from restful.request_types.shared import Entity, TypeConstrainedString
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 
 
 def validate_entity_use_case(entity: Entity, user: User, as_type: TypeConstrainedString | None) -> str:
