+++ src/features/export/use_cases/export_use_case.py
@@ -4,11 +4,11 @@ import zipfile
 
 from authentication.models import User
 from common.address import Address
+from common.providers.address_resolver import ResolvedAddress, resolve_address
 from common.tree.tree_node import Node
-from common.utils.resolve_address import ResolvedAddress, resolve_address
 from enums import SIMOS
 from features.export.use_cases.export_meta_use_case import export_meta_use_case
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 from storage.repositories.zip import ZipFileClient
