+++ src/features/search/use_cases/search_use_case.py
@@ -3,7 +3,7 @@ from copy import deepcopy
 from authentication.models import User
 from common.exceptions import BadRequestException
 from restful.request_types.shared import DataSourceList
-from services.document_service import DocumentService
+from services.document_service.document_service import DocumentService
 from storage.internal.data_source_repository import (
     DataSourceRepository,
     get_data_source,
