+++ src/services/document_service/delete_documents.py
